requirements.txt
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flask-expects-json==1.7.0
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
pymongo==4.10.1
python-dotenv==1.0.1
referencing==0.36.1
rpds-py==0.22.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
